PV2_549_576_577
PV2_575 - Already in master (Moved to Done)
PV2_549 - Minor change to PV2_205
PV2_576 - Updated price details for frameworks and standards. In case of framework change assessment price to 0 and added the original amount to training price. In case of standard, split the amount if assessment price was 0. Also, as pre Rohan's suggestion changed the framework/programmetype/pathway to 593/20/1
PV2-577 - Corrected the delivery periods